requirements.txt
requests
python-dotenv
flask
waitress
flask_cors
misskey.py
